# Use Blocks output buffer