--- a/setup.py
+++ b/setup.py
@@ -184,7 +184,8 @@ setup(
     zip_safe=False,
     packages=find_packages('src'),
     package_dir={'': 'src'},
-    package_data={'isal': ['*.pxd', '*.pyx', '*.pyi', 'py.typed',
+    package_data={'isal': ['*.pxd', '*.pyx', '*.pyi', 'py.typed', '*.md',
+                           'pycore_blocks_output_buffer.h',
                            # Include isa-l LICENSE and other relevant files
                            # with the binary distribution.
                            'isa-l/LICENSE', 'isa-l/README.md',