@rhpvorderman 2021-06-04
This runs slowers and breaks pypy compatibility :cry: 

Test with reading a file in chunks:
before (current develop):
```
Benchmark #1: python test_speed.py ~/test/big2.fastq.gz
  Time (mean ± σ):      3.346 s ±  0.024 s    [User: 3.273 s, System: 0.072 s]
  Range (min … max):    3.319 s …  3.382 s    5 runs
```
after
```
Benchmark #1: python test_speed.py ~/test/big2.fastq.gz
  Time (mean ± σ):      3.701 s ±  0.037 s    [User: 3.309 s, System: 0.390 s]
  Range (min … max):    3.646 s …  3.739 s    5 runs
```